Reuse stubs to allow to run individual tests
